refactored reusable functions into utils
