+++ kubefirst.go
@@ -1,9 +1,7 @@
 package main
 
 import (
-	"bufio"
 	"fmt"
-	"io"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -813,20 +811,6 @@ func updateDisplayWithLogs(kubefirstAPILogs, consoleLogs, kubefirstLogs *scrolli
 	}
 }
 
-func waitForQuit() {
-	reader := bufio.NewReader(os.Stdin)
-	for {
-		char, _, err := reader.ReadRune()
-		if err != nil {
-			fmt.Println("Error reading input:", err)
-			return
-		}
-		if char == 'q' || char == 'Q' {
-			return
-		}
-	}
-}
-
 func runServiceWithColoredLogs(serviceName, serviceDir, logsDir, timestamp string, printer *color.Color, cmdCreator func(string) *exec.Cmd, logs *scrollingLog) {
 	logFileName := fmt.Sprintf("%s-%s.log", serviceName, timestamp)
 	logFile := filepath.Join(logsDir, logFileName)
@@ -865,144 +849,6 @@ func runServiceWithColoredLogs(serviceName, serviceDir, logsDir, timestamp strin
 	if err != nil {
 		log.Error("Service exited with error", "service", serviceName, "error", err)
 	}
-}
-
-func logOutput(serviceName string, reader io.Reader, logFile *os.File, printer *color.Color, logs *scrollingLog) {
-	scanner := bufio.NewScanner(reader)
-	for scanner.Scan() {
-		line := scanner.Text()
-		timestamp := time.Now().Format("15:04:05")
-		formattedLine := fmt.Sprintf("[%s] %s: %s", timestamp, printer.Sprint(serviceName), line)
-		logFile.WriteString(formattedLine + "\n")
-		logs.add(formattedLine)
-	}
-}
-
-func runCommand(cmd *exec.Cmd, dir, logFile string) error {
-	cmd.Dir = dir
-	output, err := cmd.CombinedOutput()
-	if err != nil {
-		return fmt.Errorf("command failed: %w\nOutput: %s", err, output)
-	}
-	return appendToLog(logFile, string(output))
-}
-
-func appendToLog(logFile, content string) error {
-	f, err := os.OpenFile(logFile, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
-	if err != nil {
-		return err
-	}
-	defer f.Close()
-	_, err = f.WriteString(content + "\n")
-	return err
-}
-
-func runCommandWithLiveOutput(command, dir, logFile string) error {
-	cmd := exec.Command("bash", "-c", command)
-	cmd.Dir = dir
-
-	// Create a pipe for the command's stdout
-	stdout, err := cmd.StdoutPipe()
-	if err != nil {
-		return fmt.Errorf("error creating stdout pipe: %w", err)
-	}
-
-	// Create a pipe for the command's stderr
-	stderr, err := cmd.StderrPipe()
-	if err != nil {
-		return fmt.Errorf("error creating stderr pipe: %w", err)
-	}
-
-	// Start the command
-	err = cmd.Start()
-	if err != nil {
-		return fmt.Errorf("error starting command: %w", err)
-	}
-
-	// Create a log file
-	f, err := os.Create(logFile)
-	if err != nil {
-		return fmt.Errorf("error creating log file: %w", err)
-	}
-	defer f.Close()
-
-	// Create a multi-writer to write to both console and log file
-	multiWriter := io.MultiWriter(os.Stdout, f)
-
-	// Start a goroutine to read from stdout pipe
-	go func() {
-		scanner := bufio.NewScanner(stdout)
-		for scanner.Scan() {
-			fmt.Fprintln(multiWriter, color.CyanString("  [stdout] ")+scanner.Text())
-		}
-	}()
-
-	// Start a goroutine to read from stderr pipe
-	go func() {
-		scanner := bufio.NewScanner(stderr)
-		for scanner.Scan() {
-			fmt.Fprintln(multiWriter, color.RedString("  [stderr] ")+scanner.Text())
-		}
-	}()
-
-	// Wait for the command to finish
-	err = cmd.Wait()
-	if err != nil {
-		return fmt.Errorf("command failed: %w", err)
-	}
-
-	return nil
-}
-
-func runAndLogCommand(cmd *exec.Cmd, logFile string, textColor color.Attribute) error {
-	f, err := os.Create(logFile)
-	if err != nil {
-		return fmt.Errorf("error creating log file: %w", err)
-	}
-	defer f.Close()
-
-	// Create a pipe for capturing the command's output
-	r, w := io.Pipe()
-
-	// Set up a multi-writer for both the log file and the pipe
-	cmd.Stdout = io.MultiWriter(f, w)
-	cmd.Stderr = io.MultiWriter(f, os.Stderr)
-
-	colorPrinter := color.New(textColor)
-
-	// Start the command
-	err = cmd.Start()
-	if err != nil {
-		return fmt.Errorf("error starting command: %w", err)
-	}
-
-	// Read from the pipe and print colored output
-	go func() {
-		scanner := bufio.NewScanner(r)
-		for scanner.Scan() {
-			colorPrinter.Println(scanner.Text())
-		}
-		r.Close()
-	}()
-
-	// Wait for the command to finish
-	err = cmd.Wait()
-	if err != nil {
-		return fmt.Errorf("error running command: %w", err)
-	}
-
-	w.Close()
-
-	return nil
-}
-
-func getCurrentBranch(repoPath string) (string, error) {
-	cmd := exec.Command("git", "-C", repoPath, "rev-parse", "--abbrev-ref", "HEAD")
-	output, err := cmd.Output()
-	if err != nil {
-		return "", fmt.Errorf("error getting current branch: %w", err)
-	}
-	return strings.TrimSpace(string(output)), nil
 }
 
 func syncRepository(repoPath, branch string) string {
